# api/docs/swagger.yaml
definitions:
...
  healthAnalytics.AllWearableData:
    properties:
      wearableData:
        items:
          $ref: '#/definitions/healthAnalytics.WearableDataResponse'
        type: array
    type: object
...
  healthAnalytics.Lifestyles:
    properties:
      dataType:
        type: string
      dataValue:
        type: string
      recordedData:
        type: string
      userId:
        type: string
    type: object
...
  healthAnalytics.MedicalRecords:
    properties:
      attachments:
        items:
          type: string
        type: array
      description:
        type: string
      doctorId:
        type: string
      recordDate:
        type: string
      recordType:
        type: string
      userId:
        type: string
    type: object
  healthAnalytics.Message:
    properties:
      message:
        type: string
    type: object
  healthAnalytics.Monitoring:
    properties:
      Message:
        type: string
      lifestyle:
        items:
          $ref: '#/definitions/healthAnalytics.Lifestyles'
        type: array
      medRecords:
        items:
          $ref: '#/definitions/healthAnalytics.MedicalRecords'
        type: array
      wearableData:
        items:
          $ref: '#/definitions/healthAnalytics.WearableDates'
        type: array
    type: object
  healthAnalytics.MonitoringRealTime:
    properties:
      Message:
        type: string
      lifestyle:
        $ref: '#/definitions/healthAnalytics.Lifestyles'
      medRecords:
        $ref: '#/definitions/healthAnalytics.MedicalRecords'
      wearableData:
        $ref: '#/definitions/healthAnalytics.WearableDates'
    type: object
...
  healthAnalytics.UpdateWearableDate:
    properties:
      dataType:
        type: string
      dataValue:
        type: string
      deviceType:
        type: string
      id:
        type: string
    type: object
...
  healthAnalytics.WearableDataResponse:
    properties:
      createAt:
        type: string
      dataType:
        type: string
      dataValue:
        type: string
      deviceType:
        type: string
      id:
        type: string
      recordedTime:
        type: string
      updateAt:
        type: string
      userId:
        type: string
    type: object
  healthAnalytics.WearableDate:
    properties:
      dataType:
        type: string
      dataValue:
        type: string
      deviceType:
        type: string
      recordedTime:
        type: string
      userId:
        type: string
    type: object
  healthAnalytics.WearableDates:
    properties:
      dataType:
        type: string
      dataValue:
        type: string
      deviceType:
        type: string
      recordedTime:
        type: string
      userId:
        type: string
    type: object
  models.Error:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: API-Gateway service
  version: "1.0"
paths:
...
  /api/monitoring/daily-summary:
    get:
      consumes:
      - application/json
      description: Retrieves a summary of daily health monitoring data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.Monitoring'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get Daily Health Summary
      tags:
      - Monitoring
  /api/monitoring/realtime:
    get:
      consumes:
      - application/json
      description: Retrieves real-time health monitoring data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.MonitoringRealTime'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get Real-Time Health Monitoring
      tags:
      - Monitoring
  /api/monitoring/weekly-summary:
    get:
      consumes:
      - application/json
      description: Retrieves a summary of weekly health monitoring data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.Monitoring'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get Weekly Health Summary
      tags:
      - Monitoring
...
  /api/wearable-data:
    get:
      consumes:
      - application/json
      description: Lists all wearable data with optional filters
      parameters:
      - description: Pagination limit
        in: query
        name: limit
        type: string
      - description: Pagination offset
        in: query
        name: offset
        type: string
      - description: Device Type filter
        in: query
        name: device_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.AllWearableData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: List Wearable Data
      tags:
      - WearableData
    post:
      consumes:
      - application/json
      description: Adds new wearable data
      parameters:
      - description: Wearable Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/healthAnalytics.WearableDate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.WearableDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Add Wearable Data
      tags:
      - WearableData
  /api/wearable-data/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes wearable data by ID
      parameters:
      - description: Wearable Data ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete Wearable Data
      tags:
      - WearableData
    get:
      consumes:
      - application/json
      description: Retrieves wearable data by ID
      parameters:
      - description: Wearable Data ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.WearableDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get Wearable Data
      tags:
      - WearableData
    put:
      consumes:
      - application/json
      description: Updates existing wearable data by ID
      parameters:
      - description: Wearable Data ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Wearable Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/healthAnalytics.UpdateWearableDate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthAnalytics.WearableDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Update Wearable Data
      tags:
      - WearableData
schemes:
- http
swagger: "2.0"